pyproject.toml
[project]
name = "ai-chatbot"
version = "0.1.0"
description = ""
authors = [
    {name = "SultanovIlyas0@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "aiogram (>=3.21.0,<4.0.0)",
    "psycopg2 (>=2.9.10,<3.0.0)",
    "sqlalchemy (>=2.0.42,<3.0.0)",
    "python-dotenv (>=1.1.1,<2.0.0)",
    "asyncpg (>=0.30.0,<0.31.0)",
    "openai (>=1.98.0,<2.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
